30/5 - Packing State FIXED
Fixed an issue where when the bot reached the plastic bag holder, it already spawned the groceries object.

To ensure that it doesn't switch states early (CollectionState handles the spawning), I added condition related to the ShopBot's NavMeshAgent velocity and put a very small value to ensure that when it stops completely before switching states.