1/6 - Implemented new feature in Feedback State
After giving feedback, the shop bot will go back to the bot stand (I just created one).

This way, the shop bot won't just stay anywhere if it is in a supermarket.

However, if on its way back someone decides to interact with it, it will interrupt this process.

This allows for people to interact with the shop bot whenever someone has already finished using it.

GOALS:
-Optimize Cart State (More specifically, display of items in the cart)

After finishing that, I should be pretty much done. Anything else would be enhancements.